+++ novae/model.py
@@ -464,15 +464,17 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
     def assign_domains(
         self,
         adata: AnnData | list[AnnData] | None = None,
-        level: int | None = 7,
+        resolution: float | None = 0.5,
+        level: int | None = None,
         n_domains: int | None = None,
-        resolution: float | None = None,
         key_added: str | None = None,
     ) -> str:
         """Assign a domain to each cell based on the "leaves" classes.
         It either (i) uses a specific `level` of the hierarchical tree,
         (ii) enforces a precise number of `n_domains`,
-        or (iii) uses the Leiden clustering with a specific `resolution`.
+        or (iii) uses the Leiden clustering on the prototypes with a specific `resolution`.
+
+        By default, uses Leiden clustering on the prototypes. To get spatial domains hierachies, use option (i) or (ii).
 
         Note:
             You'll need to run [novae.Novae.compute_representations][] first.
@@ -481,9 +483,9 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
 
         Args:
             adata: An `AnnData` object, or a list of `AnnData` objects. Optional if the model was initialized with `adata`.
+            resolution: Resolution for the Leiden clustering.
             level: Level of the domains hierarchical tree (i.e., number of different domains to assigned).
             n_domains: If `level` is not providing the desired number of domains, use this argument to enforce a precise number of domains.
-            resolution: Resolution for the Leiden clustering. If `None`, uses the hierarchical clustering instead.
             key_added: The spatial domains will be saved in `adata.obs[key_added]`. By default, it is `adata.obs["novae_domains_X]`, where `X` is the `level` argument.
 
         Returns:
@@ -495,7 +497,9 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
             f"Did not found `adata.obs['{Keys.LEAVES}']`. Please run `model.compute_representations(...)` first"
         )
 
-        if resolution is not None:
+        if level is None and n_domains is None:
+            assert resolution is not None, "Provide at least one of `level`, `resolution`, or `n_domains`."
+
             _leiden_codes = self._leiden_prototypes(resolution=resolution)
 
             key_added = f"{Keys.DOMAINS_PREFIX}res{resolution}"
@@ -508,6 +512,8 @@ class Novae(L.LightningModule, PyTorchModelHubMixin):
             return key_added
 
         if n_domains is not None:
+            assert level is None, "Provide either `level` or `n_domains`, not both."
+
             leaves_indices = utils.unique_leaves_indices(adatas)
             level = self.swav_head.find_level(leaves_indices, n_domains)
             return self.assign_domains(adatas, level=level, key_added=key_added)