+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Starting from this version, it will be as backward compatible as possible.
 - Move `novae.utils.load_dataset` to `novae.load_dataset`
 - Move `novae.utils.quantile_scaling` to `novae.data.quantile_scaling`
 - Move `novae.utils.toy_dataset` to `novae.data.toy_dataset`
+- Use `resolution` instead of `level` by default in `assign_domains`
 - Migrate to `uv` + `ruff`
 - Move representations to `numpy` when torch is not needed anymore